Time: 55 ms (90.57%), Space: 45.9 MB (35.2%) - LeetHub

+++ 0943-sum-of-subarray-minimums/0943-sum-of-subarray-minimums.cpp
@@ -34,9 +34,9 @@ public:
     int sumSubarrayMins(vector<int>& arr) {
         vector<int>r=nsee(arr);
         vector<int>l=pse(arr);
-        long long ans=0;
+        long ans=0;
         for(int i=0;i<arr.size();i++){
-            long left=i-l[i];
+            long long left=i-l[i];
             long right=r[i]-i;
             ans+=(right*left)*arr[i]%mod;
             ans%=mod;
